# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My-Advantage App",
        "description": "",
        "contact": {
            "email": "solomon.ahamba@botosoft.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:85/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "api/v1": {},
        "/customers/register": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Create new customers",
                "description": "Create new customers",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "firstname": "a3fb6",
                                    "lastname": "Jessica Smith",
                                    "email": "solomon.ahamba@botosoft.com",
                                    "phone_number": 9034426192,
                                    "password": "password",
                                    "password_confirmation": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A Customer Created Successfully"
                    },
                    "422": {
                        "description": "A Customer Creation Validation Error"
                    },
                    "400": {
                        "description": "A Customer Creation Error"
                    },
                    "419": {
                        "description": "A Customer 419"
                    },
                    "405": {
                        "description": "A Customer Creation Method Not Allowed"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/customers/login": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer Login",
                "description": "Customer Login",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "solomon.ahamba@botosoft.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A Customer Login Successfully"
                    },
                    "422": {
                        "description": "A Customer Logining In Validation Error"
                    },
                    "400": {
                        "description": "A Customer Logining Error"
                    },
                    "419": {
                        "description": "A Customer 419"
                    },
                    "405": {
                        "description": "A Customer Creation Method Not Allowed"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/customers/profile": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get the list of resources",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "Return a list of resources"
                    },
                    "401": {
                        "description": "Authorization required."
                    },
                    "500": {
                        "description": "Server Error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Customer",
            "description": "Customer Endpoints "
        }
    ],
    "security": [
        []
    ]
}